refactor(engine): extract wiring tests to sibling files
primer-engine/src/wiring.rs carried seven contiguous trailing test modules
(classifier/extractor/comprehension construction, qnn dispatch, main-backend
plan, build-main-backend, resolve-fallback-model), pushing it to 1672 lines.
Move each into its own sibling file under src/wiring/, declared from wiring.rs
as `#[cfg(test)] mod <name>;`.

Splitting into seven separate sibling files (rather than nesting all seven
under one wiring/tests.rs) keeps every `use super::*` resolving to `wiring`
verbatim — no glob-of-glob trap, zero visibility churn, no signature change,
no edits to moved code. wiring.rs drops 1672 -> 827 lines (still over the ~500
guideline; the residue is production code, a later production-split concern).
All 45 wiring tests preserved at their exact module paths and passing.

Verified: cargo test -p primer-engine --lib (72 ok), cargo clippy
-p primer-engine --all-targets -D warnings clean, cargo fmt --all --check
clean, cargo test --workspace (51 ok blocks, 0 failures).
