+++ src/crates/primer-engine/src/wiring.rs
@@ -812,861 +812,16 @@ pub async fn build_openai_compat_embedder(
 }
 
 #[cfg(test)]
-mod classifier_construction_tests {
-    use super::*;
-
-    /// Build a minimal `BackendParams` for testing.
-    fn params_with(
-        classifier_backend: Option<&str>,
-        classifier_model: Option<&str>,
-    ) -> BackendParams {
-        BackendParams {
-            api_key: None,
-            ollama_url: "http://localhost:11434".into(),
-            openai_compat_url: "http://localhost:8000".into(),
-            openai_compat_api_key: None,
-            classifier_backend: classifier_backend.map(String::from),
-            classifier_model: classifier_model.map(String::from),
-            extractor_backend: None,
-            extractor_model: None,
-            comprehension_backend: None,
-            comprehension_model: None,
-            qnn_bundle_dir: None,
-            qnn_qairt_lib_dir: None,
-            gguf_path: None,
-            llamacpp_gpu_layers: None,
-            llamacpp_n_ctx: None,
-            reasoning_markers: Vec::new(),
-            fallback_backend: None,
-            fallback_model: None,
-            router_mode: primer_core::router::RouterMode::LocalOnly,
-            primary_ttft_budget_ms: None,
-        }
-    }
-
-    /// With main=stub and no classifier flags, we get a stub classifier.
-    #[tokio::test]
-    async fn stub_main_no_flags_gives_stub_classifier() {
-        let params = params_with(None, None);
-        let main = build_backend("stub", "main".into(), &params).await.unwrap();
-        let c = build_classifier(main, "stub", "main", &params, ClassifierSettings::default())
-            .await
-            .unwrap();
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    /// `--classifier-backend stub` wins regardless of the main backend.
-    #[tokio::test]
-    async fn explicit_stub_override_wins_over_any_main() {
-        // Main is stub (would default to stub anyway), but the explicit flag
-        // should also work when the main is declared as a different type.
-        // We use stub here to avoid needing a real cloud backend in tests.
-        let params_main = params_with(None, None);
-        let main = build_backend("stub", "main-model".into(), &params_main)
-            .await
-            .unwrap();
-        let params = params_with(Some("stub"), None);
-        let c = build_classifier(
-            main,
-            "cloud",
-            "main-model",
-            &params,
-            ClassifierSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    /// No classifier flags on a non-stub main → LlmEngagementClassifier wrapping
-    /// Arc::clone of main (identifier is "llm:<main-model>").
-    #[tokio::test]
-    async fn non_stub_main_no_flags_gives_llm_with_main_model() {
-        // We simulate a non-stub main by using "stub" type but treating it as
-        // cloud in the dispatch (the type doesn't matter for the Arc::clone path;
-        // what matters is that main_backend_name is NOT "stub").
-        // Since we can't construct a real cloud backend without an API key,
-        // we build a stub backend but pass "cloud" as the name to exercise
-        // the dispatch arm that reuses main via Arc::clone.
-        let params = params_with(None, None);
-        let main = build_backend("stub", "claude-sonnet-4-6".into(), &params)
-            .await
-            .unwrap();
-        let c = build_classifier(
-            main,
-            "cloud", // use "cloud" as name to exercise the non-stub arm
-            "claude-sonnet-4-6",
-            &params,
-            ClassifierSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "llm:claude-sonnet-4-6");
-    }
-
-    /// `--classifier-backend ollama` without `--classifier-model` is an error.
-    #[tokio::test]
-    async fn classifier_backend_without_model_is_error() {
-        let params_main = params_with(None, None);
-        let main = build_backend("stub", "main".into(), &params_main)
-            .await
-            .unwrap();
-        let params = params_with(Some("ollama"), None /* no model */);
-        let result =
-            build_classifier(main, "stub", "main", &params, ClassifierSettings::default()).await;
-        assert!(
-            result.is_err(),
-            "should error when --classifier-backend is non-stub and --classifier-model is missing"
-        );
-    }
-
-    /// `--classifier-model` override on a stub main still yields stub classifier
-    /// (the stub arm fires before we reach the model-override arm).
-    #[tokio::test]
-    async fn classifier_model_override_on_stub_main_still_yields_stub() {
-        let params = params_with(None, Some("override-model"));
-        let main = build_backend("stub", "main".into(), &params).await.unwrap();
-        let c = build_classifier(main, "stub", "main", &params, ClassifierSettings::default())
-            .await
-            .unwrap();
-        // Main is stub → classifier defaults to stub regardless of model override.
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    /// `--classifier-model` override on a NON-stub main constructs a fresh
-    /// backend of the same type with the override model (case 5 in the dispatch
-    /// matrix).
-    ///
-    /// The `main_backend` Arc is a stub (we cannot unit-test real cloud/ollama
-    /// without live infrastructure), but `main_backend_name` is "ollama" so the
-    /// dispatch falls through to the model-override branch and calls
-    /// `build_backend("ollama", "override-model", params)`.  `OllamaBackend::new`
-    /// is purely constructive (no I/O), so this succeeds in unit tests.
-    #[tokio::test]
-    async fn classifier_model_override_on_non_stub_main_constructs_fresh_backend() {
-        let params = params_with(None, Some("override-model"));
-        // The Arc itself is a stub, but the *name* string "ollama" drives dispatch.
-        let main = build_backend("stub", "main".into(), &params).await.unwrap();
-        let c = build_classifier(
-            main,
-            "ollama", // non-stub name → falls through to model-override arm
-            "main-model",
-            &params,
-            ClassifierSettings::default(),
-        )
-        .await
-        .unwrap();
-        // A fresh OllamaBackend is constructed with "override-model"; the
-        // classifier identifier must reflect the override, not the main model.
-        assert_eq!(c.identifier(), "llm:override-model");
-    }
-
-    /// An unknown `--classifier-backend` value must return an error (the
-    /// explicit-backend arm calls `build_backend`, which errors on unknown names).
-    #[tokio::test]
-    async fn unknown_classifier_backend_returns_error() {
-        let params = params_with(Some("nonexistent"), Some("any-model"));
-        let main = build_backend("stub", "main".into(), &params).await.unwrap();
-        let result =
-            build_classifier(main, "stub", "main", &params, ClassifierSettings::default()).await;
-        assert!(
-            result.is_err(),
-            "should error when --classifier-backend names an unknown backend"
-        );
-    }
-}
-
+mod build_main_backend_tests;
 #[cfg(test)]
-mod extractor_construction_tests {
-    use super::*;
-
-    fn params(extractor_backend: Option<&str>, extractor_model: Option<&str>) -> BackendParams {
-        BackendParams {
-            api_key: Some("k".into()),
-            ollama_url: "http://localhost:11434".into(),
-            openai_compat_url: "http://localhost:8000".into(),
-            openai_compat_api_key: None,
-            classifier_backend: None,
-            classifier_model: None,
-            extractor_backend: extractor_backend.map(String::from),
-            extractor_model: extractor_model.map(String::from),
-            comprehension_backend: None,
-            comprehension_model: None,
-            qnn_bundle_dir: None,
-            qnn_qairt_lib_dir: None,
-            gguf_path: None,
-            llamacpp_gpu_layers: None,
-            llamacpp_n_ctx: None,
-            reasoning_markers: Vec::new(),
-            fallback_backend: None,
-            fallback_model: None,
-            router_mode: primer_core::router::RouterMode::LocalOnly,
-            primary_ttft_budget_ms: None,
-        }
-    }
-
-    fn stub_main_backend() -> Arc<dyn InferenceBackend> {
-        Arc::new(primer_inference::stub::StubBackend)
-    }
-
-    #[tokio::test]
-    async fn stub_main_no_flags_gives_stub_extractor() {
-        let e = build_extractor(
-            stub_main_backend(),
-            "stub",
-            "stub",
-            &params(None, None),
-            ExtractorSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(e.identifier(), "stub");
-    }
-
-    #[tokio::test]
-    async fn explicit_stub_override_wins() {
-        let e = build_extractor(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(Some("stub"), None),
-            ExtractorSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(e.identifier(), "stub");
-    }
-
-    #[tokio::test]
-    async fn nonstub_backend_without_model_errors() {
-        let r = build_extractor(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(Some("ollama"), None),
-            ExtractorSettings::default(),
-        )
-        .await;
-        assert!(r.is_err());
-    }
-
-    // Note: `stub_main_backend()` always returns a StubBackend (whose
-    // `name()` is "stub"), even when `main_backend_name` says otherwise.
-    // The identifier therefore reads as `llm:stub:<model>` in these
-    // tests — what they actually verify is that the *model* name flows
-    // through correctly. In production the main_backend_name and the
-    // backend's own `name()` agree by construction.
-    #[tokio::test]
-    async fn no_flags_reuses_main_model_id() {
-        let e = build_extractor(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(None, None),
-            ExtractorSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(e.identifier(), "llm:stub:llama3.2");
-    }
-
-    #[tokio::test]
-    async fn model_only_override_uses_main_backend_type() {
-        // With `--extractor-model haiku` and main backend `ollama`,
-        // `build_backend("ollama", "haiku", ...)` constructs a real
-        // OllamaBackend, so the identifier carries the real backend
-        // name (unlike the `_reuses_main_model_id` test above which
-        // Arc::clones the stub harness).
-        let e = build_extractor(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(None, Some("haiku")),
-            ExtractorSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(e.identifier(), "llm:ollama:haiku");
-    }
-}
-
+mod classifier_construction_tests;
 #[cfg(test)]
-mod comprehension_construction_tests {
-    use super::*;
-
-    fn params(
-        comprehension_backend: Option<&str>,
-        comprehension_model: Option<&str>,
-    ) -> BackendParams {
-        BackendParams {
-            api_key: Some("k".into()),
-            ollama_url: "http://localhost:11434".into(),
-            openai_compat_url: "http://localhost:8000".into(),
-            openai_compat_api_key: None,
-            classifier_backend: None,
-            classifier_model: None,
-            extractor_backend: None,
-            extractor_model: None,
-            comprehension_backend: comprehension_backend.map(String::from),
-            comprehension_model: comprehension_model.map(String::from),
-            qnn_bundle_dir: None,
-            qnn_qairt_lib_dir: None,
-            gguf_path: None,
-            llamacpp_gpu_layers: None,
-            llamacpp_n_ctx: None,
-            reasoning_markers: Vec::new(),
-            fallback_backend: None,
-            fallback_model: None,
-            router_mode: primer_core::router::RouterMode::LocalOnly,
-            primary_ttft_budget_ms: None,
-        }
-    }
-
-    fn stub_main_backend() -> Arc<dyn InferenceBackend> {
-        Arc::new(primer_inference::stub::StubBackend)
-    }
-
-    #[tokio::test]
-    async fn stub_main_no_flags_gives_stub_comprehension() {
-        let c = build_comprehension(
-            stub_main_backend(),
-            "stub",
-            "stub",
-            &params(None, None),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    #[tokio::test]
-    async fn explicit_stub_override_wins() {
-        let c = build_comprehension(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(Some("stub"), None),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    #[tokio::test]
-    async fn nonstub_backend_without_model_errors() {
-        let r = build_comprehension(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(Some("ollama"), None),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await;
-        assert!(r.is_err());
-    }
-
-    #[tokio::test]
-    async fn no_flags_reuses_main_model_id() {
-        let c = build_comprehension(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(None, None),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "llm:stub:llama3.2");
-    }
-
-    #[tokio::test]
-    async fn model_only_override_uses_main_backend_type() {
-        let c = build_comprehension(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(None, Some("haiku")),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "llm:ollama:haiku");
-    }
-}
-
+mod comprehension_construction_tests;
 #[cfg(test)]
-mod qnn_dispatch_tests {
-    //! Pin the `--backend qnn` dispatch path of [`build_backend`].
-    //!
-    //! The QNN backend's *positive* construction path is unit-tested
-    //! at the [`primer_inference::qnn`] module (via the mock
-    //! `GenieLibrary` trait split). Here we only need to pin
-    //! [`build_backend`]'s dispatch:
-    //!
-    //! - With the feature compiled in, missing `qnn_bundle_dir` is a
-    //!   clear error from `build_backend` itself (before any FFI).
-    //! - With the feature compiled in and `qnn_bundle_dir` set,
-    //!   dispatch reaches `QnnBackend::new`, which on every non-Android
-    //!   host returns the typed `PlatformUnsupported` error — proving
-    //!   that the qnn arm fired (any other arm would have produced
-    //!   either Ok or a different error message).
-    //! - With the feature *not* compiled in, the user gets a
-    //!   "rebuild with --features qnn" hint, NOT the generic
-    //!   "unknown backend" message — that distinction is the
-    //!   load-bearing UX win of the per-feature dispatch.
-    use super::*;
-
-    /// Build a `BackendParams` skeleton for the qnn dispatch tests.
-    fn params() -> BackendParams {
-        BackendParams {
-            api_key: None,
-            ollama_url: "http://localhost:11434".into(),
-            openai_compat_url: "http://localhost:8000".into(),
-            openai_compat_api_key: None,
-            classifier_backend: None,
-            classifier_model: None,
-            extractor_backend: None,
-            extractor_model: None,
-            comprehension_backend: None,
-            comprehension_model: None,
-            qnn_bundle_dir: None,
-            qnn_qairt_lib_dir: None,
-            gguf_path: None,
-            llamacpp_gpu_layers: None,
-            llamacpp_n_ctx: None,
-            reasoning_markers: Vec::new(),
-            fallback_backend: None,
-            fallback_model: None,
-            router_mode: primer_core::router::RouterMode::LocalOnly,
-            primary_ttft_budget_ms: None,
-        }
-    }
-
-    #[test]
-    fn default_qairt_lib_dir_lives_one_dir_up_alongside_qairt_lib() {
-        // The conventional QAIRT layout from AI Hub apps puts the
-        // bundle dir alongside `qairt/` at the same parent level:
-        //   ~/primer-bundles/qwen3-4b/{genie_config.json, ...}
-        //   ~/primer-bundles/qairt/lib/aarch64-android/libGenie.so
-        let bundle = PathBuf::from("/home/user/primer-bundles/qwen3-4b");
-        let lib = default_qairt_lib_dir(&bundle);
-        assert_eq!(
-            lib,
-            PathBuf::from("/home/user/primer-bundles/qairt/lib/aarch64-android")
-        );
-    }
-
-    #[test]
-    fn default_qairt_lib_dir_tolerates_root_bundle_path() {
-        // A bundle at the filesystem root (no parent) is unusual but
-        // should not panic. We fall back to a same-directory relative
-        // path — downstream `RealGenieLibrary::open` will report a
-        // useful `LibraryLoad` error rather than us deciding here.
-        let bundle = PathBuf::from("/");
-        let lib = default_qairt_lib_dir(&bundle);
-        // On `/`, `.parent()` is `None`, so we fall back to the bare
-        // relative path. Exact form is documented in the helper.
-        assert_eq!(lib, PathBuf::from("qairt/lib/aarch64-android"));
-    }
-
-    #[test]
-    fn resolve_qairt_lib_dir_passes_explicit_through_on_every_platform() {
-        // An explicit override always wins, regardless of platform — the
-        // user knows exactly where their QAIRT libs are.
-        let bundle = PathBuf::from("/home/user/bundles/qwen3-4b");
-        let explicit = PathBuf::from("/opt/qairt/lib/aarch64-android");
-        for is_android in [false, true] {
-            assert_eq!(
-                resolve_qairt_lib_dir_for(Some(explicit.clone()), &bundle, is_android),
-                explicit,
-                "explicit override must pass through (is_android={is_android})"
-            );
-        }
-    }
-
-    #[test]
-    fn resolve_qairt_lib_dir_android_absent_yields_empty_for_basename_load() {
-        // On Android the 9 QAIRT `.so`s ship inside the APK's
-        // `lib/arm64-v8a/` (extracted to nativeLibraryDir). An absent
-        // override must resolve to an EMPTY path, which signals the QNN
-        // backend to dlopen `libGenie.so` by basename so the system
-        // linker resolves it (and its DT_NEEDED deps) from that dir.
-        let bundle = PathBuf::from("/data/local/bundles/qwen3-4b");
-        assert_eq!(
-            resolve_qairt_lib_dir_for(None, &bundle, true),
-            PathBuf::new(),
-            "Android + no override must yield empty (basename load)"
-        );
-    }
-
-    #[test]
-    fn resolve_qairt_lib_dir_desktop_absent_falls_back_to_bundle_relative() {
-        // On desktop an absent override keeps today's behaviour: the
-        // conventional `<bundle>/../qairt/lib/aarch64-android` layout.
-        let bundle = PathBuf::from("/home/user/bundles/qwen3-4b");
-        assert_eq!(
-            resolve_qairt_lib_dir_for(None, &bundle, false),
-            default_qairt_lib_dir(&bundle),
-            "desktop + no override must use the bundle-relative default"
-        );
-    }
-
-    /// Without the `qnn` feature, the dispatch arm hands back a build
-    /// hint, NOT the generic "unknown backend: qnn" string. The
-    /// distinction matters because users who haven't compiled in qnn
-    /// need a different action than users who typo'd the backend name.
-    #[cfg(not(feature = "qnn"))]
-    #[tokio::test]
-    async fn qnn_without_feature_returns_build_hint() {
-        let p = params();
-        let result = build_backend("qnn", "qnn-placeholder".into(), &p).await;
-        let err = match result {
-            Ok(_) => panic!("expected qnn-without-feature to error, got Ok"),
-            Err(e) => e,
-        };
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("qnn") && msg.contains("feature"),
-            "expected build hint mentioning qnn + feature; got: {msg}"
-        );
-        // And NOT the generic unknown-backend phrasing:
-        assert!(
-            !msg.contains("unknown backend"),
-            "qnn-without-feature should be distinct from unknown-backend; got: {msg}"
-        );
-    }
-
-    /// With the `qnn` feature compiled in but no `qnn_bundle_dir`
-    /// set in params, the dispatch arm reports the missing required
-    /// input BEFORE any FFI is attempted — exactly the "fast clap-style
-    /// rejection" UX the plan calls for.
-    #[cfg(feature = "qnn")]
-    #[tokio::test]
-    async fn qnn_with_feature_missing_bundle_dir_errors_pre_ffi() {
-        let p = params();
-        let result = build_backend("qnn", "qnn-placeholder".into(), &p).await;
-        let err = match result {
-            Ok(_) => panic!("expected qnn-with-no-bundle to error, got Ok"),
-            Err(e) => e,
-        };
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("qnn-bundle-dir") || msg.contains("bundle"),
-            "expected missing-bundle-dir hint; got: {msg}"
-        );
-    }
-
-    /// With the `qnn` feature compiled in and a `qnn_bundle_dir`
-    /// set, dispatch reaches `QnnBackend::new`. On every host the
-    /// repo's CI runs on (Linux + macOS), this surfaces the typed
-    /// `PlatformUnsupported` error from `primer-qnn-sys`. That proves
-    /// the qnn arm fired — neither the "unknown backend" arm nor the
-    /// "missing bundle dir" guard could have produced this string.
-    #[cfg(all(feature = "qnn", not(target_os = "android")))]
-    #[tokio::test]
-    async fn qnn_with_feature_and_bundle_dir_dispatches_to_real_lib_on_host() {
-        // Build params with a fake (nonexistent) bundle dir — the
-        // dispatch arm hands these straight to `QnnBackend::new`, which
-        // tries to dlopen `libGenie.so` from the qairt lib dir FIRST.
-        // On a non-Android host, that returns `PlatformUnsupported`
-        // before the bundle's existence is checked.
-        let p = BackendParams {
-            qnn_bundle_dir: Some(PathBuf::from("/nonexistent/bundle")),
-            ..params()
-        };
-        let result = build_backend("qnn", "qnn-placeholder".into(), &p).await;
-        let err = match result {
-            Ok(_) => panic!("expected PlatformUnsupported on non-Android host, got Ok"),
-            Err(e) => e,
-        };
-        let msg = format!("{err}");
-        // The dev-facing error string from
-        // `GenieCallError::PlatformUnsupported` carries the platform
-        // name. On macOS this is `"macos"`, on Linux it's `"linux"`.
-        assert!(
-            msg.to_lowercase().contains("android")
-                || msg.to_lowercase().contains("platform")
-                || msg.to_lowercase().contains("only supported"),
-            "expected PlatformUnsupported-flavoured error; got: {msg}"
-        );
-    }
-
-    /// Without the `llamacpp` feature (the default test build), the
-    /// dispatch arm hands back a build hint mentioning `llamacpp` and
-    /// `feature` — not the generic "unknown backend" string.
-    #[tokio::test]
-    async fn llamacpp_without_feature_returns_build_hint() {
-        let params = BackendParams {
-            gguf_path: Some(std::path::PathBuf::from("/tmp/model.gguf")),
-            ..params()
-        };
-        let err = match build_backend("llamacpp", "ignored".into(), &params).await {
-            Ok(_) => panic!("expected llamacpp-without-feature to error, got Ok"),
-            Err(e) => e,
-        };
-        let msg = format!("{err}");
-        assert!(msg.contains("llamacpp"), "got: {msg}");
-        assert!(msg.contains("feature"), "got: {msg}");
-    }
-
-    /// The `"llamacpp"` string reaches its own dispatch arm (the error
-    /// mentions llamacpp, proving it didn't fall through to the generic
-    /// unknown-backend arm).
-    #[tokio::test]
-    async fn llamacpp_dispatch_reaches_arm() {
-        let params = params();
-        let err = match build_backend("llamacpp", "ignored".into(), &params).await {
-            Ok(_) => panic!("expected llamacpp dispatch to error, got Ok"),
-            Err(e) => e,
-        };
-        assert!(format!("{err}").to_lowercase().contains("llamacpp"));
-    }
-}
-
+mod extractor_construction_tests;
 #[cfg(test)]
-mod main_backend_plan_tests {
-    use super::{MainBackendPlan, plan_main_backend};
-
-    #[test]
-    fn primary_ok_no_fallback_is_primary_alone() {
-        assert_eq!(
-            plan_main_backend(true, false, false),
-            MainBackendPlan::PrimaryAlone
-        );
-        assert_eq!(
-            plan_main_backend(true, false, true),
-            MainBackendPlan::PrimaryAlone
-        );
-    }
-
-    #[test]
-    fn primary_ok_fallback_built_is_wrapped() {
-        assert_eq!(
-            plan_main_backend(true, true, true),
-            MainBackendPlan::Wrapped
-        );
-    }
-
-    #[test]
-    fn primary_ok_fallback_failed_is_primary_alone() {
-        assert_eq!(
-            plan_main_backend(true, true, false),
-            MainBackendPlan::PrimaryAlone
-        );
-    }
-
-    #[test]
-    fn primary_failed_secondary_built_is_secondary_alone() {
-        assert_eq!(
-            plan_main_backend(false, true, true),
-            MainBackendPlan::SecondaryAlone
-        );
-    }
-
-    #[test]
-    fn primary_failed_secondary_failed_is_fail() {
-        assert_eq!(plan_main_backend(false, true, false), MainBackendPlan::Fail);
-    }
-
-    #[test]
-    fn primary_failed_no_fallback_is_fail() {
-        assert_eq!(
-            plan_main_backend(false, false, false),
-            MainBackendPlan::Fail
-        );
-        assert_eq!(plan_main_backend(false, false, true), MainBackendPlan::Fail);
-    }
-}
-
+mod main_backend_plan_tests;
 #[cfg(test)]
-mod build_main_backend_tests {
-    use super::*;
-    use primer_core::router::RouterMode;
-
-    fn params(
-        fallback_backend: Option<&str>,
-        fallback_model: Option<&str>,
-        router_mode: primer_core::router::RouterMode,
-    ) -> BackendParams {
-        BackendParams {
-            api_key: None,
-            ollama_url: "http://localhost:11434".into(),
-            openai_compat_url: "http://localhost:8000".into(),
-            openai_compat_api_key: None,
-            classifier_backend: None,
-            classifier_model: None,
-            extractor_backend: None,
-            extractor_model: None,
-            comprehension_backend: None,
-            comprehension_model: None,
-            qnn_bundle_dir: None,
-            qnn_qairt_lib_dir: None,
-            gguf_path: None,
-            llamacpp_gpu_layers: None,
-            llamacpp_n_ctx: None,
-            reasoning_markers: Vec::new(),
-            fallback_backend: fallback_backend.map(String::from),
-            fallback_model: fallback_model.map(String::from),
-            router_mode,
-            primary_ttft_budget_ms: None,
-        }
-    }
-
-    /// No fallback configured ⇒ primary alone (unchanged behavior).
-    #[tokio::test]
-    async fn no_fallback_returns_primary() {
-        let p = params(None, None, RouterMode::LocalOnly);
-        let b = build_main_backend("stub", "m".into(), &p).await.unwrap();
-        assert_eq!(b.name(), "stub");
-    }
-
-    /// Primary fails to build, fallback (stub) builds ⇒ secondary alone.
-    /// `unknown-backend` is an unbuildable backend name, so the primary leg errors.
-    #[tokio::test]
-    async fn primary_unbuildable_falls_back_to_secondary() {
-        let p = params(Some("stub"), None, RouterMode::LocalOnly);
-        let b = build_main_backend("unknown-backend", "m".into(), &p)
-            .await
-            .unwrap();
-        // Secondary stub served ⇒ its name surfaces.
-        assert_eq!(b.name(), "stub");
-    }
-
-    /// Primary builds, fallback fails to build ⇒ primary alone, no error.
-    #[tokio::test]
-    async fn fallback_unbuildable_keeps_primary() {
-        let p = params(Some("unknown-backend"), Some("m"), RouterMode::LocalOnly);
-        let b = build_main_backend("stub", "m".into(), &p).await.unwrap();
-        assert_eq!(b.name(), "stub");
-    }
-
-    /// Both unbuildable ⇒ error (the primary's construction error).
-    #[tokio::test]
-    async fn both_unbuildable_errors() {
-        let p = params(Some("unknown-fallback"), Some("m"), RouterMode::LocalOnly);
-        let r = build_main_backend("unknown-primary", "m".into(), &p).await;
-        // `Result::err` drops the `Ok` value (`Arc<dyn InferenceBackend>` is
-        // not `Debug`, so `expect_err` won't compile here).
-        let err = r.err().expect("both legs unbuildable must error");
-        // Spec invariant: the `Fail` arm surfaces the PRIMARY's (most
-        // informative) error, not the secondary's. Distinct backend names let
-        // us prove which one propagated.
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("unknown-primary"),
-            "expected the primary's error to surface; got: {msg}"
-        );
-        assert!(
-            !msg.contains("unknown-fallback"),
-            "must not surface the secondary's error; got: {msg}"
-        );
-    }
-
-    /// Fallback misconfigured (ollama without a model) but the primary is
-    /// healthy ⇒ the broken opt-in fallback is dropped and the primary serves
-    /// alone. A misconfigured fallback must NEVER abort startup when the
-    /// primary built fine.
-    #[tokio::test]
-    async fn fallback_misconfigured_keeps_primary() {
-        let p = params(Some("ollama"), None, RouterMode::LocalOnly);
-        let b = build_main_backend("stub", "m".into(), &p).await.unwrap();
-        assert_eq!(b.name(), "stub");
-    }
-
-    /// Primary unbuildable AND fallback misconfigured ⇒ error. The `Fail` arm
-    /// surfaces the PRIMARY's (more actionable) error, not the fallback's
-    /// resolve message.
-    #[tokio::test]
-    async fn primary_unbuildable_and_fallback_misconfigured_errors() {
-        // `ollama` without a model is a resolve error (misconfigured fallback).
-        let p = params(Some("ollama"), None, RouterMode::LocalOnly);
-        let r = build_main_backend("unknown-primary", "m".into(), &p).await;
-        let err = r.err().expect("both legs unusable must error");
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("unknown-primary"),
-            "expected the primary's error to surface; got: {msg}"
-        );
-    }
-
-    /// Hybrid mode + a buildable secondary ⇒ a RouterBackend whose name() is
-    /// the primary's (load-bearing for the small-context budget).
-    #[tokio::test]
-    async fn hybrid_with_secondary_builds_router_named_after_primary() {
-        let p = params(Some("stub"), None, RouterMode::Hybrid);
-        let b = build_main_backend("stub", "m".into(), &p).await.unwrap();
-        assert_eq!(b.name(), "stub");
-    }
-
-    /// Routing mode with NO secondary configured ⇒ a clear error.
-    #[tokio::test]
-    async fn hybrid_without_secondary_errors() {
-        let p = params(None, None, RouterMode::Hybrid);
-        let r = build_main_backend("stub", "m".into(), &p).await;
-        let err = r.err().expect("routing without a secondary must error");
-        let msg = format!("{err}").to_lowercase();
-        assert!(
-            msg.contains("secondary") || msg.contains("fallback"),
-            "expected a 'needs a secondary leg' error; got: {err}"
-        );
-    }
-
-    /// local-only mode is byte-for-byte the existing fallback behavior.
-    #[tokio::test]
-    async fn local_only_with_fallback_is_unchanged_fallback() {
-        let p = params(Some("stub"), None, RouterMode::LocalOnly);
-        let b = build_main_backend("stub", "m".into(), &p).await.unwrap();
-        assert_eq!(b.name(), "stub");
-    }
-}
-
+mod qnn_dispatch_tests;
 #[cfg(test)]
-mod resolve_fallback_model_tests {
-    use super::resolve_fallback_model;
-    use primer_core::consts::inference::DEFAULT_CLOUD_MODEL;
-
-    #[test]
-    fn stub_ignores_model() {
-        assert_eq!(resolve_fallback_model("stub", None).unwrap(), None);
-        assert_eq!(
-            resolve_fallback_model("stub", Some("x".into())).unwrap(),
-            None
-        );
-    }
-
-    #[test]
-    fn cloud_defaults_when_unset() {
-        assert_eq!(
-            resolve_fallback_model("cloud", None).unwrap(),
-            Some(DEFAULT_CLOUD_MODEL.to_string())
-        );
-    }
-
-    #[test]
-    fn cloud_uses_explicit_model() {
-        assert_eq!(
-            resolve_fallback_model("cloud", Some("claude-opus-4-7".into())).unwrap(),
-            Some("claude-opus-4-7".to_string())
-        );
-    }
-
-    #[test]
-    fn ollama_requires_model() {
-        assert!(resolve_fallback_model("ollama", None).is_err());
-        assert_eq!(
-            resolve_fallback_model("ollama", Some("llama3.2".into())).unwrap(),
-            Some("llama3.2".to_string())
-        );
-    }
-
-    #[test]
-    fn openai_compat_requires_model() {
-        assert!(resolve_fallback_model("openai-compat", None).is_err());
-        assert_eq!(
-            resolve_fallback_model("openai-compat", Some("Qwen3-8B".into())).unwrap(),
-            Some("Qwen3-8B".to_string())
-        );
-    }
-}
+mod resolve_fallback_model_tests;
